tools/selftest/resctrl: Enhance MBA test to compare the real bandwidth
Bytedance reported that using MBA might not get the linear result.
For example, if the schemata is set to no limit (100%), we get
the max bandwidth m1. But when we shrottle the MBA further, say,
set it to 60%, we expected to get memory bandwidth <= 0.6 x m1,
the actual result shows that this is not aligned.

Introduce this linear test to evaluate if the platform supports
linear test:

 Fail: Check MBA diff within 8% for schemata 60
 Fail: Check MBA accuracy within 8% for schemata 60 actual 2354 expected 2087 max 3479
 avg_diff_per: 98%
 avg_bw_imc: 2354
 avg_bw_resc: 4674
 avg_expected_diff_per: 12%

In above test, the max bandwith is 3479 when schemata is set to 100%.
However, after we set the schemata to 60%, we get the 2354, which
is much higher than expected 2087.

Reported-by: Jun Jin <jun.i.jin@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>